[BE] feat : 마이페이지 관련 API 수정 및 추가, 1. 마이페이지 - 로그인 유저 정보 반환 2. 마이페이지 - 로그인 유저의 응답 받은 코드, 응답 한 코드 반환

+++ backend/Aldy/src/main/java/com/example/demo/repository/EditedCodeRepository.java
@@ -10,6 +10,6 @@ public interface EditedCodeRepository extends JpaRepository<EditedCode, Long> {
 
     List<EditedCode> findByReceiver_id(long member_id);
 
-    int countBySender_id(long member_id);
-    int countByReceiver_id(long member_id);
+    Long countBySender_id(long member_id);
+    Long countByReceiver_id(long member_id);
 }
